New parameters

+++ src/haplotype.hpp
@@ -168,11 +168,11 @@ struct haplotype {
         return muts_in_range;
     }
 
-    int mutation_distance(const raw_read& other) const {
+    float mutation_distance(const raw_read& other) const {
         return ::mutation_distance(other.mutations, this->get_mutations(other.start, other.end));
     }
 
-    int mutation_distance(haplotype * other) const {
+    float mutation_distance(haplotype * other) const {
         return ::mutation_distance(other->get_mutations(), this->get_mutations());
     }
 
@@ -190,7 +190,7 @@ struct multi_haplotype {
     /* arena indices of ranged node children */
     std::vector<int> children;
 
-    int mutation_distance(std::vector<mutation> const& comp, int min_pos, int max_pos) {
+    float mutation_distance(std::vector<mutation> const& comp, int min_pos, int max_pos) {
         return ::mutation_distance(comp, root->get_mutations(min_pos, max_pos));
     }
 };
